% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_test_example}
\alias{bulk_test_example}
\title{Bulk test example}
\format{
A data frame with 58581 rows and 5 variables
}
\usage{
bulk_test_example
}
\description{
Bulk test example
}
\examples{
data(bulk_test_example)
}
\keyword{datasets}
