% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_cell.R
\name{calculate_avg_exp}
\alias{calculate_avg_exp}
\title{Calculate bootstrap mean expression for single cell data}
\usage{
calculate_avg_exp(raw, n_bootstrap, seed)
}
\arguments{
\item{raw}{data frame of single cell data with celltype or cluster as colnames}

\item{n_bootstrap}{number of bootstrap}

\item{seed}{random seed}
}
\value{
mean expression data
}
\description{
Calculate bootstrap mean expression for single cell data
}
