% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_iso_score}
\alias{calculate_iso_score}
\title{Calculate reaction score for isoenzymes (or relationship only)}
\usage{
calculate_iso_score(x, data, list, gene_num)
}
\arguments{
\item{x}{index}

\item{data}{input gene expression data}

\item{list}{isoenzymes reaction list}

\item{gene_num}{file of number of times one gene has involved in all pathways}
}
\value{
Calculated isoenzyme list scores
}
\description{
Calculate reaction score for isoenzymes (or relationship only)
}
