% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_multi_comp}
\alias{calculate_multi_comp}
\title{calculate reaction score for complicated complex(with both 'and' and 'or' relationship)}
\usage{
calculate_multi_comp(x, data, gene_num)
}
\arguments{
\item{x}{index}

\item{data}{input gene expression data}

\item{gene_num}{file of number of times one gene has involved in all pathways}
}
\value{
Calculated complicated enzyme complex list scores
}
\description{
calculate reaction score for complicated complex(with both 'and' and 'or' relationship)
}
