% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_reaction_score}
\alias{calculate_reaction_score}
\title{Calculate metabolic reaction scores (MRAS) for 13082 reactions}
\usage{
calculate_reaction_score(data)
}
\arguments{
\item{data}{gene expression data.1.The gene expression matrix should be gene by sample matrix where row names are human gene names (gene symbols),
and column names should be sample names. Please note that METAFlux does not support other gene IDs.
2.The input gene expression matrix should be normalized (e.g., log-transformed, etc.) before using METAFlux.
METAflux will not perform any normalization on expression data.
3.Gene expression data cannot have negative values.}
}
\value{

}
\description{
Calculate metabolic reaction scores (MRAS) for 13082 reactions
}
\examples{
 calculate_reaction_score(bulk_test_example)
}
