% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_simple_comeplex_score}
\alias{calculate_simple_comeplex_score}
\title{Calculate reaction score for enzyme complex(and relationship only)}
\usage{
calculate_simple_comeplex_score(x, data, list, gene_num)
}
\arguments{
\item{x}{index}

\item{data}{input gene expression data}

\item{list}{enzyme complex reaction list}

\item{gene_num}{file of number of times one gene has involved in all pathways}
}
\value{
Calculated enzyme complex list scores
}
\description{
Calculate reaction score for enzyme complex(and relationship only)
}
