% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_medium}
\alias{cell_medium}
\title{This file contains the hams medium  nutrients information}
\format{
A data frame with 44 rows and 2 variables:
\describe{
\item{metabolite}{stands for name of nutrients available in the medium. }
\item{reaction_name}{V2 stands for reaction ID of corresponding nutreints}

}
}
\usage{
cell_medium
}
\description{
This default file indicates the nutrients available in the medium.
There are total 44 metabolites available. This file can be changed depending on users.
We will allow users to define the medium profile based on their knowledge.
}
\examples{
data(cell_medium)
}
\keyword{datasets}
