% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scflux_calculation.R
\name{compute_sc_flux}
\alias{compute_sc_flux}
\title{sc-rna seq flux calculation}
\usage{
compute_sc_flux(num_cell, fraction, fluxscore, medium)
}
\arguments{
\item{num_cell}{the number of cell types or clusters}

\item{fraction}{fraction of each cell types. Fractions need to sum up to 1}

\item{fluxscore}{calculated metabolic activity score(MRAS)}

\item{medium}{Medium profile}
}
\value{
Flux score for single cell data
}
\description{
sc-rna seq flux calculation
}
