% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_cell.R
\name{generate_boots}
\alias{generate_boots}
\title{Generate bootstrap index data}
\usage{
generate_boots(celltype, n)
}
\arguments{
\item{celltype}{colnames of single cell data.Colnames should be labeled as cell type or cluster.}

\item{n}{number of bootstrap}
}
\value{

}
\description{
Generate bootstrap index data
}
