% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_cell.R
\name{get_ave_exp}
\alias{get_ave_exp}
\title{Calculate mean expression for one bootstrap}
\usage{
get_ave_exp(i, raw, samples)
}
\arguments{
\item{i}{index}

\item{raw}{single cell data}

\item{samples}{generated bootstrap index data}
}
\value{

}
\description{
Calculate mean expression for one bootstrap
}
