% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gem}
\alias{human_gem}
\title{Excel data of human gem model}
\format{
A data frame with 13082 rows and 16 variables:
\describe{
\item{ID}{Reaction ID}
\item{NAME}{Name of reaction}
\item{EQUATION}{Equation of reaction}
...
}
}
\source{
\url{https://github.com/SysBioChalmers/Human-GEM/blob/main/model/Human-GEM.xlsx}
}
\usage{
human_gem
}
\description{
Contains the reaction, gene, gene protein relationship,pathway information.
}
\examples{
data(human_gem)
}
\keyword{datasets}
