% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrient_lookup_files}
\alias{nutrient_lookup_files}
\title{1648 exchange reactions:mathematical representation of uptake/secrete
metabolites into the extracellular space}
\format{
1648 rows
}
\usage{
nutrient_lookup_files
}
\description{
1648 exchange reactions:mathematical representation of uptake/secrete
metabolites into the extracellular space
}
\keyword{datasets}
