% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_test_example}
\alias{sc_test_example}
\title{single cell test seurat object example}
\format{
A toy example containing 350 cells(Tumor and T cells)
}
\usage{
sc_test_example
}
\description{
single cell test seurat object example
}
\examples{
data(sc_test_example)
}
\keyword{datasets}
