% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_cell.R
\name{calculate_avg_exp}
\alias{calculate_avg_exp}
\title{Calculate bootstrap mean expression for single cell data}
\usage{
calculate_avg_exp(myseurat, myident, n_bootstrap, seed)
}
\arguments{
\item{myseurat}{Seurat object. METAFlux will calculate on "data" slot}

\item{myident}{Seurat idents for grouping.This will be a character string indicating the grouping of the seurat object}

\item{n_bootstrap}{number of bootstrap}

\item{seed}{random seed}
}
\value{
mean expression data
}
\description{
Calculate bootstrap mean expression for single cell data
}
