% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_cell.R
\name{get_ave_exp}
\alias{get_ave_exp}
\title{Calculate mean expression for one bootstrap}
\usage{
get_ave_exp(i, myseurat, samples, myident)
}
\arguments{
\item{i}{index}

\item{myseurat}{single cell Seurat object.METAFlux will calculate on "data" slot}

\item{samples}{generated bootstrap index data}

\item{myident}{Seurat idents.This will be a character string indicating the grouping of the seurat object}
}
\value{

}
\description{
Calculate mean expression for one bootstrap
}
